server/requirements.txt
Flask
Flask-Cors
SpeechRecognition
pyttsx3
python-dotenv
requests
beautifulsoup4
